+++ packages/node/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- fix get runtimeVersion failed when fetch service initialization
+- set useDictionary to false when one of the event/extrinsic filters are not provided
 
 ## [0.17.1] - 2021-06-29
 Upgrade priority: High.